# fame-server/swagger.json
{"swagger":"2.0","info":{"description":"Documentation for Blog API","version":"1.0.0","title":"Blog API Documentation","termsOfService":"https://github.com/missaouib/Blog","contact":{"name":"zzzzbw","url":"https://www.raylian.fr/blog","email":"zzzzbw@gmail.com"}},"host":"localhost","basePath":"/","tags":[{"name":"auth-controller","description":"Auth Controller"},{"name":"category-controller","description":"Category Controller"},{"name":"comment-controller","description":"Comment Controller"},{"name":"front-controller","description":"Front Controller"},{"name":"media-controller","description":"Media Controller"},{"name":"note-controller","description":"Note Controller"},{"name":"option-controller","description":"Option Controller"},{"name":"post-controller","description":"Post Controller"},{"name":"tag-controller","description":"Tag Controller"}],"paths":{"/api/admin/category":{"get":{"tags":["category-controller"],"summary":"getAll","operationId":"getAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«MetaInfo»»"}}},"deprecated":false},"post":{"tags":["category-controller"],"summary":"save","operationId":"saveUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Category»"}}},"deprecated":false},"delete":{"tags":["category-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Category»"}}},"deprecated":false}},"/api/admin/category/{id}":{"post":{"tags":["category-controller"],"summary":"update","operationId":"updateUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"},{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Category»"}}},"deprecated":false}},"/api/admin/comment":{"get":{"tags":["comment-controller"],"summary":"index","operationId":"indexUsingGET","produces":["*/*"],"parameters":[{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Comment»»"}}},"deprecated":false}},"/api/admin/comment/count":{"get":{"tags":["comment-controller"],"summary":"count","operationId":"countUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«long»"}}},"deprecated":false}},"/api/admin/comment/{id}":{"get":{"tags":["comment-controller"],"summary":"detail","operationId":"detailUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Comment»"}}},"deprecated":false},"delete":{"tags":["comment-controller"],"summary":"delete","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/login":{"post":{"tags":["auth-controller"],"summary":"login","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"password","in":"query","description":"password","required":true,"type":"string"},{"name":"rememberMe","in":"query","description":"rememberMe","required":false,"type":"string"},{"name":"username","in":"query","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/logout":{"post":{"tags":["auth-controller"],"summary":"logout","operationId":"logoutUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/media":{"get":{"tags":["media-controller"],"summary":"index","operationId":"indexUsingGET_1","produces":["*/*"],"parameters":[{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Media»»"}}},"deprecated":false}},"/api/admin/media/upload":{"post":{"tags":["media-controller"],"summary":"upload","operationId":"uploadUsingPOST","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"},{"name":"path","in":"query","description":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Media»"}}},"deprecated":false}},"/api/admin/media/{id}":{"get":{"tags":["media-controller"],"summary":"detail","operationId":"detailUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Media»"}}},"deprecated":false},"delete":{"tags":["media-controller"],"summary":"delete","operationId":"deleteUsingDELETE_2","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/note":{"get":{"tags":["note-controller"],"summary":"page","operationId":"pageUsingGET","produces":["*/*"],"parameters":[{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"},{"name":"priority","in":"query","required":false,"type":"integer","format":"int32"},{"name":"status","in":"query","required":false,"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},{"name":"title","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Note»»"}}},"deprecated":false},"post":{"tags":["note-controller"],"summary":"save","operationId":"saveUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"allowComment","in":"query","description":"allowComment","required":false,"type":"boolean","default":false},{"name":"content","in":"query","description":"content","required":true,"type":"string"},{"name":"id","in":"query","description":"id","required":false,"type":"integer","format":"int32"},{"name":"priority","in":"query","description":"priority","required":false,"type":"integer","default":0,"format":"int32"},{"name":"status","in":"query","description":"status","required":true,"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«int»"}}},"deprecated":false}},"/api/admin/note/count":{"get":{"tags":["note-controller"],"summary":"count","operationId":"countUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«long»"}}},"deprecated":false}},"/api/admin/note/{id}":{"get":{"tags":["note-controller"],"summary":"get","operationId":"getUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Note»"}}},"deprecated":false},"delete":{"tags":["note-controller"],"summary":"delete","operationId":"deleteUsingDELETE_3","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Note»"}}},"deprecated":false}},"/api/admin/option/all":{"get":{"tags":["option-controller"],"summary":"getAllOptions","operationId":"getAllOptionsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Map«string,string»»"}}},"deprecated":false}},"/api/admin/option/save":{"post":{"tags":["option-controller"],"summary":"saveAllOptions","operationId":"saveAllOptionsUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"options","in":"query","description":"options","required":true,"items":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/post":{"get":{"tags":["post-controller"],"summary":"page","operationId":"pageUsingGET_1","produces":["*/*"],"parameters":[{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"},{"name":"priority","in":"query","required":false,"type":"integer","format":"int32"},{"name":"status","in":"query","required":false,"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},{"name":"title","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Post»»"}}},"deprecated":false},"post":{"tags":["post-controller"],"summary":"save","operationId":"saveUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"allowComment","in":"query","description":"allowComment","required":false,"type":"boolean","default":false},{"name":"category","in":"query","description":"category","required":true,"type":"string"},{"name":"content","in":"query","description":"content","required":true,"type":"string"},{"name":"created","in":"query","description":"created","required":true,"type":"integer","format":"int64"},{"name":"id","in":"query","description":"id","required":false,"type":"integer","format":"int32"},{"name":"modified","in":"query","description":"modified","required":true,"type":"integer","format":"int64"},{"name":"priority","in":"query","description":"priority","required":false,"type":"integer","default":0,"format":"int32"},{"name":"status","in":"query","description":"status","required":true,"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},{"name":"tags","in":"query","description":"tags","required":true,"type":"string"},{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«int»"}}},"deprecated":false}},"/api/admin/post/count":{"get":{"tags":["post-controller"],"summary":"count","operationId":"countUsingGET_2","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«long»"}}},"deprecated":false}},"/api/admin/post/{id}":{"get":{"tags":["post-controller"],"summary":"get","operationId":"getUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Post»"}}},"deprecated":false},"delete":{"tags":["post-controller"],"summary":"delete","operationId":"deleteUsingDELETE_4","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Post»"}}},"deprecated":false}},"/api/admin/reset/password":{"post":{"tags":["auth-controller"],"summary":"resetPassword","operationId":"resetPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"newPassword","in":"query","description":"newPassword","required":true,"type":"string"},{"name":"oldPassword","in":"query","description":"oldPassword","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/reset/user":{"post":{"tags":["auth-controller"],"summary":"resetUser","operationId":"resetUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"email","in":"query","description":"email","required":true,"type":"string"},{"name":"username","in":"query","description":"username","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/admin/tag":{"get":{"tags":["tag-controller"],"summary":"getAll","operationId":"getAllUsingGET_1","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«MetaInfo»»"}}},"deprecated":false},"post":{"tags":["tag-controller"],"summary":"save","operationId":"saveUsingPOST_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Tag»"}}},"deprecated":false},"delete":{"tags":["tag-controller"],"summary":"delete","operationId":"deleteUsingDELETE_5","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Tag»"}}},"deprecated":false}},"/api/admin/tag/{id}":{"post":{"tags":["tag-controller"],"summary":"update","operationId":"updateUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"},{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Tag»"}}},"deprecated":false}},"/api/admin/user":{"get":{"tags":["auth-controller"],"summary":"getUser","operationId":"getUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«User»"}}},"deprecated":false}},"/api/archive":{"get":{"tags":["front-controller"],"summary":"archive","operationId":"archiveUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«Archive»»"}}},"deprecated":false}},"/api/category":{"get":{"tags":["front-controller"],"summary":"category","operationId":"categoryUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«MetaInfo»»"}}},"deprecated":false}},"/api/comment":{"get":{"tags":["front-controller"],"summary":"getArticleComment","operationId":"getArticleCommentUsingGET","produces":["*/*"],"parameters":[{"name":"articleId","in":"query","description":"articleId","required":true,"type":"integer","format":"int32"},{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Comment»»"}}},"deprecated":false},"post":{"tags":["front-controller"],"summary":"addComment","operationId":"addCommentUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"articleId","in":"query","description":"articleId","required":true,"type":"integer","format":"int32"},{"name":"content","in":"query","description":"content","required":true,"type":"string"},{"name":"email","in":"query","description":"email","required":false,"type":"string"},{"name":"name","in":"query","description":"name","required":true,"type":"string"},{"name":"parentId","in":"query","description":"parentId","required":false,"type":"integer","format":"int32"},{"name":"website","in":"query","description":"website","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/comment/{commentId}/assess":{"post":{"tags":["front-controller"],"summary":"assessComment","operationId":"assessCommentUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"assess","in":"query","description":"assess","required":true,"type":"string","enum":["AGREE","DISAGREE"]},{"name":"commentId","in":"path","description":"commentId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse"}}},"deprecated":false}},"/api/note":{"get":{"tags":["front-controller"],"summary":"noteList","operationId":"noteListUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«NoteInfo»»"}}},"deprecated":false}},"/api/note/{id}":{"get":{"tags":["front-controller"],"summary":"note","operationId":"noteUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Note»"}}},"deprecated":false}},"/api/option":{"get":{"tags":["front-controller"],"summary":"getOption","operationId":"getOptionUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Map«string,string»»"}}},"deprecated":false}},"/api/post":{"get":{"tags":["front-controller"],"summary":"home","operationId":"homeUsingGET","produces":["*/*"],"parameters":[{"name":"limit","in":"query","description":"limit","required":false,"type":"integer","default":10,"format":"int32"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":0,"format":"int32"},{"name":"sort","in":"query","description":"sort","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi","default":"id"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Pagination«Post»»"}}},"deprecated":false}},"/api/post/{id}":{"get":{"tags":["front-controller"],"summary":"post","operationId":"postUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«Post»"}}},"deprecated":false}},"/api/tag":{"get":{"tags":["front-controller"],"summary":"tag","operationId":"tagUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RestResponse«List«MetaInfo»»"}}},"deprecated":false}}},"definitions":{"Archive":{"type":"object","properties":{"postInfos":{"type":"array","items":{"$ref":"#/definitions/PostInfo"}},"year":{"type":"string"}},"title":"Archive"},"Category":{"type":"object","properties":{"created":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"modified":{"type":"string","format":"date-time"},"name":{"type":"string"}},"title":"Category"},"Comment":{"type":"object","properties":{"agent":{"type":"string"},"agree":{"type":"integer","format":"int32"},"articleId":{"type":"integer","format":"int32"},"content":{"type":"string"},"created":{"type":"string","format":"date-time"},"disagree":{"type":"integer","format":"int32"},"email":{"type":"string"},"id":{"type":"integer","format":"int32"},"ip":{"type":"string"},"modified":{"type":"string","format":"date-time"},"name":{"type":"string"},"parentId":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["NORMAL","DELETE"]},"website":{"type":"string"}},"title":"Comment"},"Map«string,string»":{"type":"object","title":"Map«string,string»","additionalProperties":{"type":"string"}},"Media":{"type":"object","properties":{"created":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"modified":{"type":"string","format":"date-time"},"name":{"type":"string"},"suffix":{"type":"string"},"thumbUrl":{"type":"string"},"url":{"type":"string"}},"title":"Media"},"MetaInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"postInfos":{"type":"array","items":{"$ref":"#/definitions/PostInfo"}}},"title":"MetaInfo"},"Note":{"type":"object","properties":{"allowComment":{"type":"boolean"},"authorId":{"type":"integer","format":"int32"},"commentCount":{"type":"integer","format":"int32"},"content":{"type":"string"},"created":{"type":"string","format":"date-time"},"hits":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"modified":{"type":"string","format":"date-time"},"priority":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},"title":{"type":"string"}},"title":"Note"},"NoteInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"priority":{"type":"integer","format":"int32"},"title":{"type":"string"}},"title":"NoteInfo"},"Pagination«Comment»":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/definitions/Comment"}},"orderBy":{"type":"string"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"pages":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int64"}},"title":"Pagination«Comment»"},"Pagination«Media»":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/definitions/Media"}},"orderBy":{"type":"string"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"pages":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int64"}},"title":"Pagination«Media»"},"Pagination«Note»":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/definitions/Note"}},"orderBy":{"type":"string"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"pages":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int64"}},"title":"Pagination«Note»"},"Pagination«Post»":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/definitions/Post"}},"orderBy":{"type":"string"},"pageNum":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"pages":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int64"}},"title":"Pagination«Post»"},"Post":{"type":"object","properties":{"allowComment":{"type":"boolean"},"authorId":{"type":"integer","format":"int32"},"category":{"type":"string"},"commentCount":{"type":"integer","format":"int32"},"content":{"type":"string"},"created":{"type":"string","format":"date-time"},"hits":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"modified":{"type":"string","format":"date-time"},"priority":{"type":"integer","format":"int32"},"status":{"type":"string","enum":["PUBLISH","DRAFT","DELETE"]},"tags":{"type":"string"},"title":{"type":"string"}},"title":"Post"},"PostInfo":{"type":"object","properties":{"created":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"title":{"type":"string"}},"title":"PostInfo"},"RestResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse"},"RestResponse«Category»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Category»"},"RestResponse«Comment»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Comment"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Comment»"},"RestResponse«List«Archive»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/definitions/Archive"}},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«List«Archive»»"},"RestResponse«List«MetaInfo»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/definitions/MetaInfo"}},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«List«MetaInfo»»"},"RestResponse«List«NoteInfo»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/definitions/NoteInfo"}},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«List«NoteInfo»»"},"RestResponse«Map«string,string»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object","additionalProperties":{"type":"string"}},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Map«string,string»»"},"RestResponse«Media»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Media"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Media»"},"RestResponse«Note»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Note"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Note»"},"RestResponse«Pagination«Comment»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Pagination«Comment»"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Pagination«Comment»»"},"RestResponse«Pagination«Media»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Pagination«Media»"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Pagination«Media»»"},"RestResponse«Pagination«Note»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Pagination«Note»"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Pagination«Note»»"},"RestResponse«Pagination«Post»»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Pagination«Post»"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Pagination«Post»»"},"RestResponse«Post»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/Post"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Post»"},"RestResponse«Tag»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«Tag»"},"RestResponse«User»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/definitions/User"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«User»"},"RestResponse«int»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"integer","format":"int32"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«int»"},"RestResponse«long»":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"integer","format":"int64"},"msg":{"type":"string"},"success":{"type":"boolean"}},"title":"RestResponse«long»"},"Tag":{"type":"object","properties":{"created":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"modified":{"type":"string","format":"date-time"},"name":{"type":"string"}},"title":"Tag"},"User":{"type":"object","properties":{"created":{"type":"string","format":"date-time"},"email":{"type":"string"},"id":{"type":"integer","format":"int32"},"logged":{"type":"string","format":"date-time"},"modified":{"type":"string","format":"date-time"},"passwordMd5":{"type":"string"},"screenName":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}